requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2.2
django-channels==0.7.0
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework-api-key==3.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
msgpack==1.1.0
oauthlib==3.2.2
openai==1.86.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
